remove dep on file-utils
Mostly it's just using .exists, which is a trivial wrapper around
fs.existsSync.  This is racy anyway!  Where we can, switch it to catch
ENOENT from fs.readFileSync directly.  In the one place where we can't
do that, use fs.existsSync directly.

file.expand was just wrapping glob, in a fairly inefficient way for
how we're using it.  Just call glob.sync directly and use lodash to
filter the results.